Add Path and PathSegment classes
